requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Werkzeug==2.3.7
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
gunicorn==21.2.0
